[MM-13991] Add channel users subcommand and deprecate channel add and remove
